+++ contracts/RewardMaker.sol
@@ -94,7 +94,7 @@ contract RewardMaker is Ownable {
     // F1 - F10: OK
     // F3: _convert is separate to save gas by only checking the 'onlyEOA' modifier once in case of convertMultiple
     // F6: There is an exploit to add lots of REWARD to the bar, run convert, then remove the REWARD again.
-    //     As the size of the StakeReward has grown, this requires large amounts of funds and isn't super profitable anymore
+    //     As the size of the SwivelStaked has grown, this requires large amounts of funds and isn't super profitable anymore
     //     The onlyEOA modifier prevents this being done with a flash loan.
     // C1 - C24: OK
     function convert(address token0, address token1) external onlyEOA() {