+++ contracts/SwivelToken.sol
@@ -5,7 +5,7 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/utils/Pausable.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
-contract RewardToken is ERC20, Pausable, Ownable {
+contract SwivelToken is ERC20, Pausable, Ownable {
 
     /// @notice The EIP-712 typehash for the contract's domain
     bytes32 public constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
@@ -13,7 +13,7 @@ contract RewardToken is ERC20, Pausable, Ownable {
     /// @notice The EIP-712 typehash for the delegation struct used by the contract
     bytes32 public constant DELEGATION_TYPEHASH = keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
 
-    constructor() ERC20("RewardToken", "REWARD") public {
+    constructor() ERC20("Swivel Token", "SVL") public {
     }
 
     function pause() public onlyOwner {