+++ deploy/MiniChefV2.js
@@ -10,12 +10,14 @@ module.exports = async function ({ ethers, deployments, getNamedAccounts }) {
   let rewardAddress;
 
   if (chainId === "31337") {
-    rewardAddress = (await deployments.get("RewardToken")).address;
+    rewardAddress = (await deployments.get("SwivelToken")).address;
   } else if (chainId in REWARD_ADDRESS) {
     rewardAddress = REWARD_ADDRESS[chainId];
   } else {
     throw Error("No REWARD!");
   }
+
+ // rewardAddress = '0x0B295A3524c7835bb8a743e2f0D743614dA8C68A';
 
   await deploy("MiniChefV2", {
     from: deployer,