+++ test/MasterChefV2.test.ts
@@ -3,12 +3,12 @@ import { assert, expect } from "chai"
 
 describe("MasterChefV2", function () {
   before(async function () {
-    await prepare(this, ["MasterChef", "RewardToken", "ERC20Mock", "MasterChefV2", "RewarderMock", "RewarderBrokenMock"])
+    await prepare(this, ["MasterChef", "SwivelToken", "ERC20Mock", "MasterChefV2", "RewarderMock", "RewarderBrokenMock"])
     await deploy(this, [["brokenRewarder", this.RewarderBrokenMock]])
   })
 
   beforeEach(async function () {
-    await deploy(this, [["reward", this.RewardToken]])
+    await deploy(this, [["reward", this.SwivelToken]])
 
     await deploy(this, [
       ["lp", this.ERC20Mock, ["LP Token", "LPT", getBigNumber(10)]],